s/requirements.txt
numpy>=1.20.3
hdfs==2.6.0
emoji>=1.4.2,<=2.0.0
nltk==3.5
wordninja==2.0.0
python-Levenshtein==0.12.0
torch>=1.8.2,<=2.0.0
metaphone==0.6.0
